fix: git_utils.is_git_repo() - respect git command output
`git rev-parse --is-inside-work-tree` can be used to test if the command
is being run from inside of git repository because it will fail if run
from outside of a repository. However, it can also run successfully but
output 'false' if run from within a directory inside of the repository
that is not part of a worktree.

This is most likely to happen if using a bare repository and the hurl
file is kept in a location outside of a worktree. Previously, attempting
to run such a hurl request threw an error in find_env_files for
attempting to concatenate local 'git_root' (nil, in this case) and no
further hurl requests could be run from neovim.

This commit allows find_env_files to continue as if not in a git repo.