fix(split): refactor filetype setting and highlight application
This commit refactors the way the filetype is set for the split buffer
in `hurl/split.lua`. Instead of conditionally setting the filetype based
on the presence of `edgy.nvim`, the filetype is now always set to
'hurl-nvim'. This change simplifies the code and ensures consistent
behavior regardless of the presence of `edgy.nvim`.